+++ changelog.md
@@ -15,6 +15,7 @@
 - Made projectiles reflected by Maestro receive player projectile modifiers (e.g., bouncing, piercing, helix movement, etc.)
 - Fixed Natascha's spin up sound playing while the game is paused
 - Changed Midas touch effect from Quarter Pounder's mastery to trigger when actually touching statues rather than just getting near them
+- Made Midas-touched projectiles receive player projectile modifiers
 
 ## 1.29.0 (2026-04-23)